google-news-check/requirements.txt
# ETF News Publisher Analysis - Python Dependencies

gnews>=0.4.3
beautifulsoup4>=4.14.0
lxml>=6.0.0
lxml-html-clean>=0.4.0
pandas>=3.0.0
newspaper3k>=0.2.8
